Converted another tuple into a class
