+++ KXlsxConverterAPI/Models/ScheduleModels/Shift.cs
@@ -12,11 +12,3 @@ public class Shift
     public DateTime? BreakTwo { get; set; } = null;
     public string OriginalPosition { get; set; } = string.Empty;
 }
-
-public class ShiftData(DateTime start, DateTime end, JobPosition jobPosition, bool isSubshift = false)
-{
-    public DateTime Start = start;
-    public DateTime End = end;
-    public JobPosition Position = jobPosition;
-    public bool IsSubshift = isSubshift;
-}
